tcp: Time to break it - replacing old files with new ones
